move usePairs into hooks
